🚧 some donation refactoring
