+++ Toggl.Foundation.Tests/MvvmCross/BaseMvvmCrossTests.cs
@@ -4,6 +4,7 @@ using Toggl.Foundation.Analytics;
 using Toggl.Foundation.DataSources;
 using Toggl.Foundation.Interactors;
 using Toggl.Foundation.MvvmCross.Services;
+using Toggl.Foundation.Services;
 using Toggl.Foundation.Shortcuts;
 using Toggl.Multivac;
 using Toggl.PrimeRadiant;
